Refactor: modularize DXF geometry and variant handling
Moved geometry, group code variant, and line width classes from EoDxfBase.h into new headers: EoDxfGeometry.h, EoDxfGroupCodeValuesVariant.h, and EoDxfLineWidths.h. Updated includes throughout the codebase to use these new headers. Moved ETYPE and SymbolTable enums to EoDxfBase.h. Made minor corrections to int16_t/int32_t usage for DXF codes. Updated project files to include new headers. This improves modularity and maintainability of the DXF library.

+++ Core Libraries/EoDxfLib/EoDxfEntities.cpp
@@ -7,6 +7,8 @@
 
 #include "EoDxfBase.h"
 #include "EoDxfEntities.h"
+#include "EoDxfGeometry.h"
+#include "EoDxfGroupCodeValuesVariant.h"
 #include "EoDxfReader.h"
 
 EoDxfEntity::EoDxfEntity(const EoDxfEntity& other)
@@ -136,7 +138,7 @@ void EoDxfEntity::ParseCode(int code, EoDxfReader* reader) {
       m_color = reader->GetInt16();
       break;
     case 370:
-      m_lineWeight = EoDxfLineWidths::dxfInt2lineWidth(reader->GetInt32());
+      m_lineWeight = EoDxfLineWidths::dxfInt2lineWidth(reader->GetInt16());
       break;
     case 48:
       m_lineTypeScale = reader->GetDouble();
